Update patch set 1
Patch Set 1: Code-Review-1

This works, but I would prefer that we make both the width and the height of the button follow the same method (and I prefer the method you are using). This would hopefully also fix the problem of the X being much further from the right of the drawer than the top (even though the top and right values are the same).

Patch-set: 1
Label: Code-Review=-1