Update patch set 4
Patch Set 4:

For menu items added via Javascript it is as simple as before (if it was
actually done somewhere):

    var mainMenu = M.require('mainMenu'),
      mainMenuSchema = new SchemaMobileWebClickTracking({}, 'MobileWebMainMenuClickTracking');

    mainMenuSchema.hijackLink($link, 'event-name');

If you were to add data models just for this, how would you handle all the menu
items generated on the server, and bind those? The solution proposed here seems
pretty simple and works for server and browser menu links.

Also, we bind data to html tags everywhere through the views, I don't think
this is any different

Patch-set: 4