Update patch set 1
Patch Set 1: Code-Review-1

Please add a comment as we may want to revisit this if we move the last modified bar to the bottom of the page.

history-link-loaded is emitted in inline HTML and history-link-loaded is binded to in lastEdited/init.js so the comment should make reference to last modified bar binds to events.

You can also remove mobile.modules as this is a dependency of mobile.oo
Once you've done this should be good to merge

Patch-set: 1
Label: Code-Review=-1