Update patch set 3
Patch Set 3:

"I thought there was common agreement that MobileFrontend is the wrong extension for Special:UserProfile. What harm is there in moving this code out of an inappropriate place? If there's interest in (another) separate extension, the Git history is available to all."

I agree. Just the order of things is wrong. That extension should be setup on the cluster first before removing it here. Personally I think we should tackle that in a reboot/re-evaluation of this feature.

Patch-set: 3