Update patch set 8
Patch Set 8:

I discussed this with Kaldari and we are being very schizophrenic in how we do this sort of thing. We already use the class_exist methods to control Nearby in the menu and 

As discussed on the other patchset with Sam and Baha, the other approach to getting Collections into this menu AND providing event logging is also extremely fragile and led to this patchset.

It seems we are all in disagreement with the right approach.
That said we need this before Thursday when we are deploying Gather to the cluster.

I think we need to talk as a team on the right approach here. Please don't let another ideological disagreement delay Gather any more.... let's resolve this bikeshed later.

Patch-set: 8
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0