Update patch set 3
Patch Set 3:

This is mind-boggling. You're willing to hobble this feature by removing images (cf. <https://gerrit.wikimedia.org/r/193017>), but outright death is too much? This feature isn't maintained and honest-to-God if I thought there was anything worth saving here, I would've said so. But the more I looked at this, the more I agreed that it made sense to either completely start over ("reboot," which you apparently also agree with...) or use an existing extension (mentioned in the commit message), for which there is currently no community support. This doesn't belong in MobileFrontend, it doesn't belong on Wikimedia wikis, you contributed to this mess, and now you're standing in the way of the people holding brooms offering to help clean it up.

Patch-set: 3