Update patch set 2
Patch Set 2:

The key difference is that I can depend on a whole bunch of deferred things and have a single response when they are resolved.  

Anyway, I'll kill this patch but I stand my ground in that EventEmitter is not the solution for everything.

Patch-set: 2
Reviewer: Robmoen <53@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0