Update patch set 10
Patch Set 10: Code-Review-1

This doesn't really address the bug it links to. This still feels suboptimal since there is still a flash of unstyled content whilst the image loads. Maybe this code simply needs to be pulled in at the top of the page. It also seems like if doing it this way you should render it in the page itself as you are adding an API request everywhere which is not used by people without JS.

Essentially you are just moving the API request from JS to PHP with no benefit to the user.

More generally I think if we are going to put effort into this we should really look to start a PageBanner extension asap since there is also development interest from the Wikivoyage community (https://phabricator.wikimedia.org/T77925)

Patch-set: 10
Label: Code-Review=-1