Update patch set 1
Patch Set 1:

Hm? Not really, it doesn't depend on echo (if you just look at the dependencies tree). That it only works with Echo is another topic and should be done where the module is loaded/added (it's already SkinMinerva loads mobile.notification only, if Echo is installed), so it can be defined here :)

Patch-set: 1