Update patch set 1
Patch Set 1: Code-Review-1

Compatibility is pretty bad as Jon pointed out http://caniuse.com/#search=hyphens but I don't think it is a problem to have better experience on more capable browsers. It seems like Apple Microsoft and Mozilla implemented it, so I would guess that it is a matter of time that chrome gets it. It doesn't seem like a feature that is only implemented by one vendor.

I'm good with this as long as we do the vendor prefixing on a mixin in minerva.mixins.less to keep the vendor prefixing contained (that's why I -1d)

Patch-set: 1
Reviewer: Jhernandez <2027@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1