Alpha, beta: Fix page save spinner
Apparently, using buttons as icon containers in alpha doesn't work
because the background image is not aligned vertically. Removing the
container and falling back to the default container fixes the problem.
Also, it doesn't make sense to use a button for a spinner. Spinners
are note meant to be clickable.

Bug: T94881
Change-Id: I8ceac6697a435564838c9635d1fdfc0f51b135ff