Update patch set 1
Patch Set 1:

After some thinking...
A better approach might be simply to rewrite this as a View.
We do need to consolidate our code with oojs ui but I really think this is hacky and going about it the wrong way. It's the View that needs to be rewritten in oojs ui for us to adopt it.

I think https://phabricator.wikimedia.org/T88559 is the least path of resistance to rewriting our entire infrastructure in OOjs ui and I know that Ed Sanders would be really supportive in us doing that. Once that's done it should be trivial to make the big switch.

Patch-set: 1