Update patch set 1
Patch Set 1: Code-Review-1

I agree with florian that if it is called like that, if it exists then it should be used for the header (effectively allowing you to add html in the header). Either that or rename it to 'skin-minerva-disable-header' or something like that.

I don't agree with the hooks though. I would prefer the variable approach for this.

Patch-set: 1
Reviewer: Jhernandez <2027@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1