Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Contrary to the bug description, OO.ui is still a dependency. I'm not sure what problem this patch is solving. The existing code is just being wrapped in a view, giving the impression that we can change the view presentation with a template and bind events to it.

Patch-set: 2