Hygiene: Panels should not automatically be appended to DOM
The panel is exhibiting the behaviour of a drawer which creates problems
when you try to insert a Panel inside another view e.g. an overlay

A Drawer is something that gets added to the bottom of the page
A Panel can be used anywhere.

Note this *might* require a change to WikiGrokDialog but I don't understand
that code well enough to know how to do this.

Bug: T96176
Change-Id: Id171118c104a872f7e2e00de57cc125645cb4207