Update patch set 3
Patch Set 3:

Jdlrobson: There currently isn't any concrete plan to phase out the mobile login page (that I'm aware of). I would also love to get rid of it, but I don't know when that's actually going to happen. In the meantime, it seems like a bad idea to have Gather code in the MF login page (and inconsistent with how we're handling PageMessageHeaders, which uses a hook). If you feel strongly that this is a better solution for the short term, I'll merge it, but it feels hacky.

Patch-set: 3