Check, if the user is blocked from editing before try to save the edit
Instead of try to save an edit and give an error message, if that fails,
check, if the user is blocked from editing just in the moment, where the
editor loads.

This change adds only a notice, thte user is still able to edit the content
and try to save it with the well known error message, that th edit couldn't
be saved, because the user is blocked.

Bug: T96257
Change-Id: I146800e176e12d70ccb93d5d5ade934bc31740b3