Rewrite mobile login page to be less hacky and closer to core version
This makes it easier for other extensions, such as Gather, to pass
their own messages to the page (without using hooks).

This removes the mobile-only login message header, which should be
implemented in the core login page if it is actually useful.

This will also simplify the transition from a custom mobile login
page to sharing the login page with core.

Bug: T95065
Change-Id: If5ec2ca2f345ab95851ecd2d658332d94a8d4270