Update patch set 7
Patch Set 7: Code-Review-1

The editor should be closed when user is blocked. The toast could be ignored - so better to bail out of the editor and show the toast to make things crystal clear.

Also various problems with message...

"Your account or IP address is blocked from editing. The block was made by Admin for the following reason: Autoblocked because your IP address has been recently used by "[[User:BlockedUser|BlockedUser]]". The reason given for BlockedUser's block is "blah". You'll not be able to safe your edit."

* Unparsed wikitext
* "safe" => "save" (although you'll want to change message based on my feedback above any way)
* I'm also not sure how links render in toasts... I suspect not well.

Patch-set: 7
Label: Code-Review=-1