Update patch set 14
Patch Set 14:

Jdlrobson: what's the status of the PageBanner extension and how it might be used to replace this?

Lead images are still in beta. We'd like 'em to be in stable. I feel that having a lead image on the screen when the page loads is preferable to one that loads after the page has loaded.

The reason that I suggested that this patch should be resurrected is that I'm not keen on waiting around for a perfect/generic solution with no interim measure.

Patch-set: 14
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0