Update patch set 3
Patch Set 3: Code-Review-1

Not like this.. not like this. This breaks a bunch of stuff.

A lot of these messages are //not// warnings. They are informative. For example clicking on the watchlist link takes you to a login page that tells you what it is. The warning box makes you look like you have done something wrong.

Compare before and after - http://imgur.com/YYwrajA,6dv09qZ
this is a broken experience.

Patch-set: 3
Label: Code-Review=-1