Update patch set 2
Patch Set 2:

I'm really concerned that the bottom of the page is becoming a dumping ground.

Historically beta has been a staging area for things that should/could be promoted at any time. although the categories feature is pretty solid I think having three buttons at the bottom of the page is unacceptable.

We have to solve this problem first. As soon as buttons spill on to a second line this looks unprofessional... :/

Florian any bright ideas in that area?

Patch-set: 2