Update patch set 1
Patch Set 1:

> Do you really mean, that we should handle this case (i can't think about a case, when this can happen, maybe if an extension changed the contentmodel for a page and isn't correctly installed anymore?). It seems to be very rarely, that this runs into an exception?

Exceptions are, by definition, exceptional. If there isn't a case where Title#getContentModel could throw an exception then the code should be removed. Until that happens though, it /can/ throw an exception, so clients of the method have to handle it.

> Btw.: The ApiQueryInfo doesn't handle this exception, too: https://github.com/wikimedia/mediawiki/blob/master/includes/api/ApiQueryInfo.php#L389

We'll have to wrap that in a try/catch block too then :)

Patch-set: 1