Improve trademark code in MobileFrontend
Show trademark symbol in footer, but not header (per Legal)

Also, wiki administrators will now have control over which symbol
is actually displayed, registered trademark or unregistered.

The changes are (mostly) backwards compatible, except that it no
longer shows a conflicting symbol in the alt attribute (which was
wrong). It uses the same symbol in both the displayed sitename and
the alt text.

Bug: T95007
Change-Id: Ibb9a398aa79dd63c08db8e8f3fe59ec58d32fcad