Update patch set 4
Patch Set 4:

I don't think you should. Instead you should move it Schema.js because it's very simple and does the same thing as beacons. Moreover, there is no need for the 'add' functionality, to stack up multiple events. If the user is leaving a page, then it's because of the last event. That event only needs to be saved. Other events that happened prior to that event should already have been sent.

Patch-set: 4