Update patch set 1
Patch Set 1:

To upgrade this repo to the latest release of jscs (which finally fixes a long-standing concurrency bug), I had to over-ride some of the new rules (which is meh, but whatever), but also some of the existing ones. This patch restores the repo to the status quo, as it clearly says.

Patch-set: 1