Explicitly define module position
Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

This change moves 'skins.minerva.icons.images' to the bottom,
as icons don't need to block page rendering. This move soft-depends
on Ie4287e17d6f298cc63f42f257b1f67ee36961b77

Bug: T97410
Change-Id: Ibb2f3dd1e6d23cf56e95e17d7d2d66523d8e2698