Update patch set 1
Patch Set 1:

> If you can test whether it's a tablet or not early on
We could do it with the DeviceDetection class in php, but it would be cache related, so the decision, if tablet styles are used or not depends on the device, that's used to render the page, that wouldn't be that good. So the only solution i actually can think about, is to check it in JavaScript :(

> you could just hide everything with JS until the tablet styles are ready.
The problem is, that the content is styled different on mobile phone and tablet :/ Maybe we should consider to chnage that, that at least the content is styled with the same css file using less variables and media-queries?

Patch-set: 1