Update patch set 1
Patch Set 1:

Well, that's precisely what we do for styles that are considered vital. The ones that apply to all devices are in the head. They block rendering of text until they're loaded. Because showing the text, completely unstyled, as soon as we can would look so ugly people would wonder if it's broken. And the visual jumpiness of going from unstyled to styled can be very distracting.

Also, consider that these styles are currently in the head, and therefore they are making *all devices* wait on them before rendering any text, not just tablets. By hiding the text the way I suggest, the performance for tablets should be pretty close to what it is now (it does add a request with its own latency, until HTTP/SPDY comes along and that's no longer a problem), but it would be improved for all other devices.

Patch-set: 1