Update patch set 13
Patch Set 13:

Hi do you have an idea on how to do this. But for now it should be done manualy by either putting it in README file or on the extension page. but I think it should be done on extension page and README. but there can be choice here.

Patch-set: 13