Update patch set 2
Patch Set 2: Code-Review-1

This exists for a reason...

#content specifically wraps user generated content
In original designs Nearby and Watchlist and various other pages on mobile were made to take up the full screen to take up all the available space - this seems to break that and I don't see any indication on the bug that this has been thought about and is desired. None of these pages are mentioned as getting changes.

The other problem with wrapping in #content is now any community created styles will run on the special page - we scoped all of the ones in English Wikipedia for example to only apply to #content

I seem to remember there were other reasons but I'd have to do detective work via git blame to dig them out.

Please before making a big change like this ask around why 'unstyledContent' exists before doing anything or at least provide solid reasoning in the commit message to make it clear you understand the impact of what you are doing.

This also causes CSS regressions in the styling of Special:Gather

We really need to get in the habit of talking more on Phabricator bugs/mailing list about architecture changes like this.

Patch-set: 2
Label: Code-Review=-1