Update patch set 15
Patch Set 15:

Btw.: If someone is wondering, the last part pf the commit message:
"This change adds only a notice, thte user is still able to edit the content
and try to save it with the well known error message, that th edit couldn't
be saved, because the user is blocked."

Isn't correct. If the user is blocked, the editor closes immediately after the content was loaded, so the user can't edit the content.

Patch-set: 15
Reviewer: Florianschmidtwelzow <890@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0