Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Minor inline about changing the `historyData` map at the source rather than rejigging it prior to rendering, which we should be working to avoid all over the shop – we've got to start somewhere, right?

I've tested this in Chrome (42.0.2311.135) as the styling change is trivial with an emulated iPhone 6 Plus, iPad 3/4, and in regular desktop mode.

I'll ping KHammerstein on Phab about whether or not the grey background should span the full width of the screen in stable

Patch-set: 1
Label: Code-Review=-1