Update patch set 2
Patch Set 2: Code-Review+1

This reduces the styles in the head a tiny bit but this change means the page visibly adapts to tablet (triggers a repaint). I'd be interested in @Gilles opinion on this whether this is an improvement or not.

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1