Update patch set 2
Patch Set 2:

Yeah, that's not the best implementation, but iirc it is by design :/

I think that's the only easy solution to doing that, the question is only: Do we want to do that? Tablet styles are "only" a nice gimmick for tablet users, to make the content much more readable, so i think it's a benefit for all users, who take the benefit of activating JS. The other question is: Do we have some data about users with a tablet and deactivated JS? How many are that? In my point of view a tablet is a relatively modern device where the most people leave JS activated, but it's only a personal opinion :)

Patch-set: 2
Label: Code-Review=0