Update patch set 2
Patch Set 2: Code-Review-1

The plan has outlined on the bug is to hijack links to show a drawer instead.
Currently the user experience is terrible for readers who might not understand a red link means "does not exist" or for readers who are colour blind.

Secondly creating a new article is going to be fairly tricky on a mobile device for a Wikimedia site :)

Showing a drawer would allow someone to edit if they want or dismiss and go back to the article.

@Florian do you want to just do that instead and fix all the issues around this in one go?

Patch-set: 2
Label: Code-Review=-1