Update patch set 5
Patch Set 5: Code-Review-1

Kaity says to change the text to make it more positive.
She has proposed:

"This page has not yet been created on {{SITENAME}}."
She says
"The page you tried to visit does not exist." sounds too much like an error message.
Let's work out the wording and get this merged.

Also note translations will take a while to come in so maybe we want to put this on beta for a week to give time to collect these or at least merge as close to the cut off point as possible?

Patch-set: 5
Label: Code-Review=-1