Update patch set 3
Patch Set 3: Code-Review-1

This adds a bunch of code to the initial page and I'm not sure why and for waht value we are doing this. I've added a comment on the Phabricator task. Let's discuss this some more before continuing... :)

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1