Message box standardisation (step 1)
Drop our alert class in favour of the classes used in core
(warningbox, errorbox, successbox).

Make it clearer the purpose of the rule for mw-dismissable-notice by
generalising it

On history and contributions page show a warningbox rather than an errorbox
since they are not errors just information.

Refactor message boxes to give a consistent experience on:
* viewing diffs with warnings
* viewing old revisions
* MobileEditor with js disabled
* history/contributions page when no items found
* special uploads page when bad username given
* saving preferences
* warning box on special pages that are not available in stable
(currently none)
* user profile when bad username given

Change-Id: I0fed10d45285a48fbc6af50dd3a3be9940ffc186