Update patch set 5
Patch Set 5: Code-Review-1

I've tested this and it correctly outputs either the alternate or canonical link but not both. However, this patch gets rid of an early return in the hook handler – which I'm not sure about actually – and I wanted to make sure that's the intention.

Patch-set: 5
Label: Code-Review=-1