Update patch set 5
Patch Set 5:

>Could you add more context to your commit message so that I can understand the reasoning behind the change?
Sure

> Have you considered that accessing a singleton object directly everywhere is roughly equivalent to accessing the global configuration variables directly?

The goal is to remove the dependency to globals in our code. The bigger image is, that a configuration can (possibly) retrieved from other places, instead of globals, too. E.g. in one of the next steps, mediawiki-core implements other ways (or an abstract way) to store configuration values and an extension (or mw core itself) adds a way to store configuration values in the database (combined with an object cache, e.g., to access them faster). So, e.g. third-parties or possibly Wikipedias too, can change the configuration of mediawiki by itself via a webinterface without handling LocalSettings.php (you maybe know this way of configuration from other platforms, such as wordpress). For more information about that, see https://www.mediawiki.org/wiki/Manual:Configuration_for_developers

And for this goal, we need to go away from using globals directly. For that we actually use GlobalVarConfig (an implementation of Config), which simply returns the value of the global you requested, but could use the database or any other store, too. It's "just" an abstraction for our way to access configuration variables :)

Patch-set: 5