Update patch set 3
Patch Set 3:

To be clear - we have no plans to merge this or fix up this patch.

We do agree this code would be better moved to a separate extension however. Please see https://phabricator.wikimedia.org/T101049 (part of an epic to refactor MobileFrontend).

If this sounds like a reasonable plan, would you mind abandoning this patchset and help us work towards that?

Patch-set: 3