Update patch set 3
Patch Set 3:

@Jon: I wouldn't have used the exact same words as MZ, but I agree with his sentiment. I don't see the point of intentionally filing duplicate bugs, it just wastes everyone's time. 

I don't know who the "we" is that you refer to, I certainly don't mind rebasing this patch as necessary. Given that you apparently agree with the rationale explained in the commit message, I'm as baffled as MZ as to why you want me to abandon this. If the code is to be moved to a separate extension, it has to be removed from MobileFrontend at some point.

In any case, the law of -2 states that this patch will eventually be merged so I don't mind waiting a bit.

Patch-set: 3
Label: Verified=0