Fix SpecialMobileDiffTest#testNames test
To generate the diff, SpecialMobileDiff#showDiff defers to
DifferenceEngine#generateContentDiffBody. This method loads the prev and
next revisions from the DB by ID directly rather than accepting them as
parameters. This behaviour is non-trivial to isolate in a test.

The SpecialMobileDiffTest#testNames test exercises how
SpecialMobileDiff#executeWhenAvailable handles parameter validation. In
doing so, however, it invokes #showDiff. Therefore, depending on the
current state of the DB, a revision may be loaded from the DB
inadvertently.

Because the #showDiff method doesn't affect the return result of
 #executeWhenAvailable we can safely stub it and re-enable the test.

Bug: T100716
Change-Id: I19df8acd50bc32f594ad10b6e33a2dd66825f176