Update patch set 4
Patch Set 4: Code-Review-1

Mz please stop using such language. It really is toxic and unnecessary and every time you do so I lose more respect for you and your opinion.

I believe the user profile serves a useful purpose for new users. The user page is a confusing experience for non-editors. There is no structure to it and it looks like a wiki page. We are limited on space in mobile and thus we use the user profile page to provide links to contributions, user talk, edit count (which I've been told during Wikimanias that people find useful in patrolling), user page and recent changes.

Maybe to put it another way - how can we prove or disprove one way or the other that it is useful and what is useful. Currently it's just a bunch of opinions against another bunch of opinions and this patch to remove it as I've pointed out is not taking the above into account and I have no plans to merge it.

Patch-set: 4