Update patch set 4
Patch Set 4:

Well, the task says "The current implementation of a UserProfile has various issues.  Various issues is not defined, while it could be clearly defined in terms of what it is missing, and in terms of the broader context of user to user interaction on mobile, which is generally not yet heavily supported. Happy to talk more about it, MZ.

Patch-set: 4
Reviewer: Moushira <2601@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0