Update patch set 10
Patch Set 10:

Those two cases where it's set, shouldn't be setting it. These getter methods are essentially abusing a global variable to store singleton cache. The configuration variable is what it is. If its not set, then the wiki doesn't have it configured. It's fine to fallback or have a default, but that doesn't belong in the global at runtime. Not having a setter in Config is by design. (A ConfigManager might have one in the future when changing values from a Special page, but not at run time). Use a private or static member to cache the computation instead.

Patch-set: 10