QA: Fix flaky notifications step by waiting on resource ready state
The module notifications link depends on hijacking the standard link via
JS. Waiting on the resource's 'ready' state ensures the hijacking has
completed.

This pattern may be useful for other heavy JS scenarios and projects. If
it works well, we should implement a general solution in mw-selenium to
make implementing these waits easier and encapsulate exceptions so that
the errors thrown are clear to developers familiar with ResourceLoader.

Change-Id: I769282bfd5324d762e64fef4bd639e943c92853a