Update patch set 2
Patch Set 2: Code-Review-1

I think it would be better to update the JavaScript in core. I really want us to shrink the size of MobileFrontend and turn it into a skin rather than bloating it further.

Please ping me in IRC and let's see what this involves.

Patch-set: 2
Label: Code-Review=-1