QA: Refactor flaky diff test to use API for initial state
Refactored the "Added and removed content" scenario to use the API for
setting up initial page edit history which should make it more reliable
than driving the browser to accomplish the same state. Reducing the
number of given steps should also make the scope of the test case more
clear.

Change-Id: I2ea13fb7cd020b3362a926d533266a75890d4b1b