QA: Fix section toggling scenarios
Fixed issue in "Respect the hash on sections" that was incidentally
exposed by the upgrade to mw-selenium 1.x. Some step definitions that
navigate to specific articles aimed to avoid redirects incorrectly used
`String#sub` instead of `String#gsub` to replace spaces.

The scenario fails if the redirect does not take place as the page is
not reloaded when the URL fragment/hash changes. It's unclear whether or
not the expected behavior is to handle such live changes to the hash via
the `hashchange` event in which case this would be a positive failure.

Change-Id: I122b79a729eaa650b54d134875b719d21567078d