Update patch set 2
Patch Set 2: Code-Review-2

This should definitely be discussed before further work on it. Do we already have a task for it? Can you link me to it? What is the advantage for a mobile (!!!) user linking the whole bunch of special pages we have (of which some doesn't render mobile friendly).

And btw.: A big list of links without any explanation isn't really otimised for mobile. Maybe hiding the groups and show only the title is a better thing (while tapping on one group opens all special pages under it).

Patch-set: 2