Update patch set 2
Patch Set 2:

And the advantage is quicker access to special pages like if you would like to access for example Special:BrokenRedirects a user would have to type it into the search bar but what if the user forgot what it is called. He/she could then search through the list of specialpages to get the name.

Patch-set: 2