Update patch set 2
Patch Set 2: Code-Review-1

WHY? :D Please, paladox, don't "just update" a dependency, because there is an update. It would be great, if you could mention the reason, why you want to upgrade the libraries, used in this extension, either in a task in phabricator or in your commit message.

Patch-set: 2
Label: Code-Review=-1