Update patch set 1
Patch Set 1:

@ESanders generally I think now would be a really really good time to try and rewrite VisualEditorOverlay (and then EditorOverlay) in pure oojs ui and move it (at least VE) to the VisualEditor extension.

I'm concerned about patches like this and no roadmap in place for doing this.

I think ideally if I pair with someone who knows oojs ui inside out this could be done in a couple of Fridays (preferably in office in person if possible). I have a good idea of the challenges on mobile side...

Patch-set: 1
Label: Code-Review=0