Update patch set 1
Patch Set 1:

I think we understand each other and agree on the goal.

> IMO each patch should solve one specific problem. And this patch is already using styles defined in EditWatchlist.

I agree in principle but I don't think we should treat tidying up as a separate problem to be solved at another time. Instead we should endeavour to implement the cleanest solution that we can – you've called me out a few times on that yourself :)

The problem that this patch solves is "Give Special:Watchlist a max-width in mobile". How then does changing the styles for Special:Watchlist and Special:History, which has already had the max-width treatment, so that they are shared not solve this in one patch?

Patch-set: 1
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0