Inherit UsercreateTemplate for User{login|create} templates to fix T102775
AntiSpoof needs addInputItem from UsercreateTemplate to work. We reuse the hook,
so we need to extend this class in our own login/create templates.

For now we use UsercreateTemplate for Usercreate as well as for Userlogin templates,
which works, because UserlocinTemplate doesn't implement any other functions except
execute(). A solution to solve this would be to get rid of the special mobile templates
and use core login page.

Bug: T102775
Change-Id: I5eaa48a794ab461eb07ecc46810d4fb34f5e8bc3