Update patch set 1
Patch Set 1: Code-Review-1

This is a behaviour change not a bug fix.
Now when you click a search result and click the back button you are taken back to the page with search loaded.

If I remember correctly  we didn't register search with OverlayManager for the reason of preventing it from showing up in browser history.

I'd urge you to try and find a simpler fix, but if we think this behaviour is broken and want to revisit I'd rather do it knowingly rather than as a bug fix.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1