Update patch set 52
Patch Set 52: Code-Review-1

I've copied this on the task too and I'd appreciate the conversation happens there rather than on Gerrit, but here is a copy and paste job for you :-):

The latest patchset loses lots of very important comments in Resources.php and other fiels such as // FIXME: Remove when cache has cleared 
This is simply unacceptable and we need to find a way to do this in the new way.

@Paladox since this patch seems to have been going on for a while (now 50+ patchsets) maybe we should talk some more on the task about moving this forward, as I'm keen for us to make some progress here and our issues to be addressed one at a time. I think the problem here is this is actually an epic task compared to switching to extension.json in a place like Gather.

If possible technically it might make sense to do this bit by bit - @legoktm is it possible to have extension.json for say... config only and keep resources loading via a php file? If so could we move this forward that way, chunk by chunk? 

This is quite frankly a terrifyingly large patch and I'd rather have more confidence we will be able to identify errors and rollback.

Patch-set: 52