Update patch set 1
Patch Set 1: Code-Review-1

Florian - please can you outline why you are doing this in a bug? I'm not sure we should be adding more complexity to the login screen. Please also make a note of the bug that is trying to remove it (https://phabricator.wikimedia.org/T74910) in an inline comment

Patch-set: 1
Label: Code-Review=-1