Update patch set 23
Patch Set 23:

> Icon: yes. I need to see how visualeditor do that their loading bar in VE.

It's a relatively easy construct:
HTML component (printed with JS):
https://github.com/wikimedia/mediawiki-extensions-VisualEditor/blob/master/modules/ve-mw/init/targets/ve.init.mw.ViewPageTarget.init.js#L34-L41

Style (CSS) component:
https://github.com/wikimedia/mediawiki-extensions-VisualEditor/blob/master/modules/ve-mw/init/styles/ve.init.mw.ViewPageTarget.init.css#L58-L79


> JS fallback: Right now we just want to measure the performance improvement on first paint of doing this. T100778 needs more thought. I'll update the commit message.

My thoughts are, that we (in fact) totally break beta + alpha mode for our no-js users. Do we have some metrics, how much no-js users we have in beta mode? We should at least (if you think we can't add a no-js workaround, yet) add a noscript tag with a short notice, that we can't load more text and the user has to switch into stable (maybe with a direct button to do this) to see the full article.

Patch-set: 23