Update patch set 9
Patch Set 9: Code-Review-1

(1 comment)

I think that subclassing makes sense here given the current layout of the code. It might be worth revisiting it to break it apart a little more so that the transformation of the API result (SearchApi#_getPageData) into Page objects (should be done by a factory) is done outside of the class.

Minor inline following on from a previous discussion that we didn't really finish.

I'm going to try and create a test harness for this now.

Patch-set: 9
Label: Code-Review=-1