Regression: Remove last modified bar padding in stable
As part of standardization .pre-content and .post-content
classes have been added in an earlier patch. Those classes
have a max-width of 1000px and horizontal padding of 3.35em.
These rules are pushing the last modified bar content from
left to the middle of the page.

Overriding those rules reverts the regression.

Bug: T103826
Change-Id: If0c857231d99b52576e8d8c4c857bc7bbeb9d4f6