Update patch set 6
Patch Set 6: Code-Review-1

Somewhat related to what Sam said my previous review seems to have been misunderstood

When I said
"It seems all that is needed is this code?:
.stable #mw-mf-last-modified.pre-content {
  max-width: none;
}"

I was really asking why isn't this patchset just that?
It seems this patch mixes bug fix and refactor..

If there are other bugs that Sam noticed it would be helpful to test one of those at a time.

Patch-set: 6
Label: Code-Review=-1