Update patch set 3
Patch Set 3: Code-Review+2

This changes the edit screen, but given that this work is so confusing right now and it's not clear what is supposed to align with what I'd rather get the notifications menu styling fixed.

The left: -1em rule continues to bother me since it makes assumptions about the icon size which is out of its control but let's revisit that at a later time.

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2