Update patch set 3
Patch Set 3:

@Jdlrobson, take a look at this screenshot: https://phab.wmfusercontent.org/file/data/herkvplcwcogvysz27fp/PHID-FILE-mh64ljde5x2yzq4w3gqn/header-04.png

As you can see the content (i.e. everything) is left aligned with the main menu icon image (not the icon itself, because the icon has a 1em left and right margins). That -1em was introduced because of this design change. And I think it's also controlled with a variable, so I wouldn't worry about something breaking if that number changes.

Patch-set: 3
Reviewer: Bmansurov <1963@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0