Update patch set 12
Patch Set 12:

Maybe some clarification for the editing team members: This change is:

a) only an a/b test, to see, if the successful edits would increase, if there are some basic features in the wikitext editor (VE isn't available on smartphones).

b) Even, if, in the future (i haven't seen any roadmap yet), VE will be enabled on smartphones, and the wikitext editor wouldn't be maintained anymore, the users can only benefit from this change, even if we remove this functionality later and point to VE as a better editor.

So, in my point of view, even if there isn't a clear plan, how mobile editing is supported in the future, this change could help users, editing on mobile phones, and help us to understand, what they really need. So, i would really welcome some feedback, and, if there's nothing to complain about, to get this merged.

Patch-set: 12
Label: Verified=0