Update patch set 2
Patch Set 2: Code-Review-1

Maybe it's just because I'm using an old version of iOS, but the feature is buggy for me. Instead of being able to swipe, I get 2 tiny little arrows. If the arrows appear above the image, clicking on them works, but if they are overlaid on top of the image, they don't work, and instead it acts as if I clicked on the image (the footer disappears and the image is enlarged to full height).

Patch-set: 2
Reviewer: Kaldari <78@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1