Update patch set 3
Patch Set 3:

Reset class won't work. The only way to override a * selector is with another * selector. We would then have a * selector setting content boz in VE, but that would break the places we actually use border box. It should be possible to just apply border box to all the places that were previously affected by it.

Patch-set: 3